fix: move ensurePath to utils
