+++ src/utils/components/DeschedulerModal/DeschedulerModal.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import produce from 'immer';
 
-import { ensurePath } from '@catalog/utils/WizardVMContext';
 import { V1VirtualMachine, V1VirtualMachineInstance } from '@kubevirt-ui/kubevirt-api/kubevirt';
 import TabModal from '@kubevirt-utils/components/TabModal/TabModal';
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import { DESCHEDULER_EVICT_LABEL } from '@kubevirt-utils/resources/vmi';
+import { ensurePath } from '@kubevirt-utils/utils/utils';
 import { Alert, AlertVariant, Checkbox, Form, FormGroup } from '@patternfly/react-core';
 
 import { ModalPendingChangesAlert } from '../PendingChanges/ModalPendingChangesAlert/ModalPendingChangesAlert';
