+++ src/utils/utils/utils.ts
@@ -71,3 +71,18 @@ export const includeFilter = (
 
   return compareData.selected?.length === 0 || compareData.selected?.includes(compareString);
 };
+
+export const ensurePath = <T extends object>(data: T, paths: string | string[]) => {
+  let current = data;
+
+  if (Array.isArray(paths)) {
+    paths.forEach((path) => ensurePath(data, path));
+  } else {
+    const keys = paths.split('.');
+
+    for (const key of keys) {
+      if (!current[key]) current[key] = {};
+      current = current[key];
+    }
+  }
+};
